python_app/requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Werkzeug==3.1.3
google-genai==1.14.0
requests==2.32.3
python-dotenv==1.1.0
